# Conflicts:
#	conanfile.py

+++ src/libs/rigging/src/script_func.cpp
@@ -5,7 +5,7 @@
 #include "sail.h"
 #include "v_s_stack.h"
 
-extern float g_fSailHoleDepend;
+extern double g_fSailHoleDepend;
 // extern float GetSailSpeed(int holeQ,int holeMax,float maxSpeed,float fSailHoleDepend);
 
 uint32_t _ShipSailState(VS_STACK *pS)
